refactor: 🚚 update path from common/utils to src/utils
