+++ apps/nextjs/src/common/index.ts
@@ -2,4 +2,3 @@ export * from './constants';
 export * from './contexts';
 export * from './hooks';
 export * from './types';
-export * from './utils';
