+++ apps/nextjs/src/components/tooltip/tooltip.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '~/common/utils/cn';
+import { cn } from '~/utils';
 import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 
 const TooltipProvider = TooltipPrimitive.Provider;
