+++ apps/nextjs/src/utils/helpers.ts
