+++ apps/nextjs/src/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './api';
 export * from './cn';
 export * from './helpers';
